feat: 로그인 여부에 따라 레이아웃 렌더링
